SndObj 8 bit playback fixed
Former-commit-id: c826be604a0c9cdf3efd294a47b850e58dd092a7

+++ SndObj-2.6.6/src/SinSyn.cpp
@@ -47,7 +47,7 @@ SinSyn::SinSyn(SinAnal* input, int maxtracks, Table* table,
   :SndObj(input, vecsize, sr){
 
   m_ptable = table;
-  m_size = m_ptable->GetLen();
+  m_size = static_cast<float>(m_ptable->GetLen());
   m_LoTWOPI = m_size/TWOPI;
 
   m_factor = m_vecsize/m_sr;
@@ -86,7 +86,7 @@ void
 SinSyn::SetTable(Table *table)
 {
   m_ptable = table;
-  m_size = m_ptable->GetLen();
+  m_size = static_cast<float>(m_ptable->GetLen());
   m_LoTWOPI = m_size/TWOPI;
   m_ratio = m_size/m_sr;
 }
